Theme: Add default templates
page.html is a blank page template mirroring the page-specific templates that are used in the new theme. This will let designers add and preview page content on main-test.
header.html overrides the parent theme's header and removes the cover template. this is only used on post & archive views, so it likely won't be seen here, but it's added just in case.