Make sure the filter is added on rest api requests
Evidently `wp_is_json_request()` isn't reliable this early in the process.

It might be better to add the template/stylesheet filtering on an early filter such as `plugins_loaded`. I wasn't able to find one that's both early enough for theme switching to work, and late enough for `wp_is_json_request()` to work too.