Bypass the sandbox when running build:patterns
This uses `wp_remote_get()` and the `CURLOPT_CONNECT_TO` curl option to ensure `yarn build:patterns` will always connect to production servers, even if `wordpress.org` is sandboxed on your local machine.

Without this, Docker will fail to connect when you're sandboxed, because it respects /etc/hosts but doesn't use the proxy, so the request is denied.