Content: Recommend MariaDB due to MySQL 8.4+ incompatibility
MySQL 8.4 and newer are incompatible with WordPress because they have
removed support for `SQL_CALC_FOUND_ROWS`. Thus the server requirement
recommendations needs to be updated to prioritize MariaDB.

To prevent users from setting up sites on incompatible database
versions, this change updates the content to:

- Recommend MariaDB as the primary database solution.
- Specify that MySQL versions 8.0 or older are compatible where MySQL is
  mentioned.
- Update general references from "PHP and MySQL" to "PHP and MariaDB" or
  "PHP and MariaDB/MySQL" for accuracy.

This change addresses the compatibility issue and reflects MariaDB's
widespread adoption within the hosting ecosystem and its alignment with
the open source ethos of the WordPress project. Continued promotion of
Oracle's MySQL and indirectly Oracle's expanding media ecosystem is not
in the interest of WordPress and the ethos of "democratize publishing".